Moved middleware for org to it's own file
